Add firebase to project
This does a couple things. First, it loads the image from the firebase
storage. It also makes a Firebase function that takes a name of an item,
gets the location of that item in the firestore database, and checks to
see if the point given is in the bounding box.